requirements.txt
pandas==1.3.1
Flask==2.0.1
streamlit==0.86.0
flasgger==0.9.5
numpy==1.21.1
Pillow==8.3.1
scikit_learn==0.24.2
gunicorn
